__len__ for LoggerAccumulatedData + tests + micro-refactoring
